Noticket: fix test that failed when database has crtificates already
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@23726 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/systemtests/src-test/org/ejbca/core/protocol/cmp/AuthenticationModulesTest.java
@@ -515,7 +515,9 @@ public class AuthenticationModulesTest extends CmpTestCase {
                 asn1InputStream.close();
             }
         } finally {
-            this.endEntityManagementSession.deleteUser(ADMIN, userName);
+            if (endEntityManagementSession.existsUser(userName)) {
+                this.endEntityManagementSession.deleteUser(ADMIN, userName);
+            }
         }
     }
